requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
psycopg2-binary==2.9.11
Pillow==12.0.0
celery==5.6.0
redis==7.1.0
gunicorn==23.0.0
